requirements.txt
flask
flask_login
mysql-connector-python
flask_session
bleach
python-dotenv
markdown
gunicorn
